+++ src/helpers/createDummyProfile.ts
@@ -1,4 +1,4 @@
-import { type SocialSource } from '@/constants/enum.js';
+import { type SocialSource, Source } from '@/constants/enum.js';
 import { type Profile, ProfileStatus } from '@/providers/types/SocialMedia.js';
 
 export function createDummyProfile(source: SocialSource) {
@@ -14,4 +14,11 @@ export function createDummyProfile(source: SocialSource) {
         status: ProfileStatus.Active,
         verified: true,
     } satisfies Profile;
+}
+
+export function createDummyProfileFromLensHandle(handle: string) {
+    return {
+        ...createDummyProfile(Source.Lens),
+        handle,
+    };
 }
