+++ src/helpers/createWagmiPublicClient.ts
@@ -1,8 +1,15 @@
+import { createLookupTableResolver } from '@masknet/shared-base';
 import { getRPCConstant } from '@masknet/web3-shared-evm';
 import { first } from 'lodash-es';
 import { type Chain, createPublicClient as createClient, http, type PublicClient } from 'viem';
+import { polygon } from 'viem/chains';
 
-import { resolvePublicProviderUrl } from '@/helpers/resolvePublicProviderUrl.js';
+const resolvePublicProviderUrl = createLookupTableResolver<number, string>(
+    {
+        [polygon.id]: 'https://polygon-rpc.com',
+    },
+    '',
+);
 
 const map = new Map<number, PublicClient>();
