+++ src/helpers/formatPercentage.ts
@@ -1,4 +1,3 @@
-import { removeTrailingZeros } from '@/helpers/removeTrailingZeros.js';
 // eslint-disable-next-line no-restricted-imports
 import { isLessThan } from '@/maskbook/packages/web3-shared/base/src/helpers/number.js';
 
@@ -9,5 +8,9 @@ export function formatPercentage(value: number) {
     if (isLessThan(value, '0.0001')) {
         return '< 0.01%';
     }
-    return `${removeTrailingZeros((value * 100).toFixed(2))}%`;
+    const percentage = (value * 100)
+        .toFixed(2)
+        .replace(/(\.\d*?)0+$/, '$1')
+        .replace(/\.$/, '');
+    return `${percentage}%`;
 }
