+++ src/helpers/getPollFrameUrl.ts
@@ -2,4 +2,6 @@ import urlcat from 'urlcat';
 
 import { FRAME_SERVER_URL } from '@/constants/index.js';
 
-export const getPollFrameUrl = (pollId: string) => urlcat(FRAME_SERVER_URL, `/polls/${pollId}`);
+export function getPollFrameUrl(pollId: string) {
+    return urlcat(FRAME_SERVER_URL, `/polls/${pollId}`);
+}
