--- src/helpers/handleZodErrorMessage.ts
@@ -1,8 +1,0 @@
-import type { ZodError } from 'zod';
-
-export function handleZodErrorMessage(error: ZodError) {
-    return (
-        'InvalidParams: ' +
-        error.issues.map((issue) => `(${issue.code})${issue.path.join('.')}: ${issue.message}`).join('; ')
-    );
-}
