--- src/helpers/removeTrailingZeros.ts
@@ -1,3 +1,0 @@
-export function removeTrailingZeros(str: string) {
-    return str.replace(/(\.\d*?)0+$/, '$1').replace(/\.$/, '');
-}
