--- src/helpers/types.ts
@@ -1,1 +1,0 @@
-export type NonUndefined<T> = T extends undefined ? never : T;
