+++ src/locales/zh-Hant/messages.po
@@ -1307,4 +1307,3 @@ msgstr ""
 
 msgid "<0/> quoted your <1><2/></1>"
 msgstr ""
-
